dashboard/requirements.txt
streamlit==1.37.1
snowflake-connector-python==3.12.3
pandas==3.0.0
